Move `exceptions` a level up
