+++ src/neptune_scale/__init__.py
@@ -4,7 +4,7 @@ __all__ = [
 
 import warnings
 
-from neptune_scale.api.exceptions import NeptuneScaleWarning
 from neptune_scale.api.run import Run
+from neptune_scale.exceptions import NeptuneScaleWarning
 
 warnings.simplefilter("once", category=NeptuneScaleWarning)
