+++ src/neptune_scale/exceptions.py
