+++ tests/unit/test_metadata_splitter.py
@@ -19,7 +19,7 @@ from neptune_api.proto.neptune_pb.ingest.v1.common_pb2 import (
 from neptune_api.proto.neptune_pb.ingest.v1.pub.ingest_pb2 import RunOperation
 from pytest import mark
 
-from neptune_scale.api.exceptions import NeptuneFloatValueNanInfUnsupported
+from neptune_scale.exceptions import NeptuneFloatValueNanInfUnsupported
 from neptune_scale.sync.metadata_splitter import MetadataSplitter
